download dialog for bucket
