+++ src/views/pages/Storage/Browser/Browser.js
@@ -17,10 +17,6 @@ import {
   Checkbox,
   TableSortLabel,
   Slide,
-  FormControl,
-  FormControlLabel,
-  FormGroup,
-  FormLabel,
 } from "@material-ui/core";
 import AddIcon from "@material-ui/icons/Add";
 import DeleteIcon from "@material-ui/icons/Delete";
@@ -30,17 +26,12 @@ import PublishIcon from "@material-ui/icons/Publish";
 import CreateNewFolderIcon from "@material-ui/icons/CreateNewFolder";
 import ShareIcon from "@material-ui/icons/Share";
 import EditIcon from "@material-ui/icons/Edit";
-import {ModalManager, TextField} from "@material-ui/core"
 import "./style.css";
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 import { useTheme } from '@material-ui/core/styles';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import GetAppIcon from "@material-ui/icons/GetApp";
 const createBucketData = (name, accessMode) => {
   return { name, accessMode };
 };
@@ -412,6 +403,11 @@ const BucketItemsContainer = ({ title, items, onBack, onItemClick }) => {
   const isMenuOpen = Boolean(anchorEl);
 
   const menuId = "mobile-menu";
+  const [openDownloadDialog, setOpenDownloadDialog] = useState(false);
+
+  const handleOpenDownload = (state) => {
+    setOpenDownloadDialog(state);
+  };
   const renderMenu = (
     <Menu
       anchorEl={anchorEl}
@@ -456,12 +452,13 @@ const BucketItemsContainer = ({ title, items, onBack, onItemClick }) => {
               <Button startIcon={<PublishIcon />}>Upload file</Button>
               <Button startIcon={<PublishIcon />}>Upload folder</Button>
               <Button startIcon={<CreateNewFolderIcon />}>Create folder</Button>
+              <Button startIcon={<GetAppIcon />}>Download</Button>
             </div>
             <div style={{ flexGrow: "1" }}></div>
             <div className="browser-appbar-button-group">
               <Button startIcon={<ShareIcon />}>Share</Button>
               <Button startIcon={<EditIcon />}>Edit bucket</Button>
-              <Button startIcon={<DeleteIcon />}>Delete</Button>
+              <Button startIcon={<DeleteIcon />} onClick={() => handleOpenDownload(true)}>Delete</Button>
             </div>
             <div className="browser-appbar-mobile-menu">
               <IconButton
@@ -483,9 +480,72 @@ const BucketItemsContainer = ({ title, items, onBack, onItemClick }) => {
           items={items || []}
           onItemClick={onItemClick}
         />
+        <ConfirmDownload open={openDownloadDialog}/>
         {renderMenu}
       </Paper>
     </Slide>
+  );
+};
+
+const ConfirmDownload = ({open}) => {
+  const [openDownloadDialog, setOpenDownloadDialog] = React.useState(open);
+
+  const handleOpenDownload = (state) => {
+    setOpenDownloadDialog(state);
+  };
+
+  return (
+    <div>
+    <ToastContainer/>
+      {openDownloadDialog ? (
+          <>
+            <div className="justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none">
+              <div className="relative w-auto my-6 mx-auto max-w-3xl">
+                {/*content*/}
+                <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
+                  {/*header*/}
+                  <div className="flex items-start justify-between p-5 border-b border-solid border-gray-300 rounded-t">
+                    <h3 className="text-3xl font-semibold">
+                      Download these files?
+                    </h3>
+                    <button
+                        className="p-1 ml-auto bg-transparent border-0 text-black opacity-5 float-right text-3xl leading-none font-semibold outline-none focus:outline-none"
+                        onClick={() => handleOpenDownload(false)}
+                    >
+                  <span className="bg-transparent text-black opacity-5 h-6 w-6 text-2xl block outline-none focus:outline-none">
+                    ×
+                  </span>
+                    </button>
+                  </div>
+                  {/*body*/}
+                  <div className="relative p-6 flex-auto">
+
+                  </div>
+                  {/*footer*/}
+                  <div className="flex items-center justify-end p-6 border-t border-solid border-gray-300 rounded-b">
+                    <button
+                        className="text-red-500 background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1"
+                        type="button"
+                        style={{ transition: "all .15s ease" }}
+                        onClick={() => handleOpenDownload(false)}
+                    >
+                      Cancel
+                    </button>
+                    <button
+                        className="bg-light-blue text-white active:bg-light-blue font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1"
+                        type="button"
+                        style={{ transition: "all .15s ease" }}
+                        onClick={() => handleOpenDownload(false)}
+                    >
+                      OK
+                    </button>
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
+          </>) : null}
+    </div>
   );
 };
 
@@ -548,50 +608,11 @@ const BucketItemTable = ({
 
   const emptyRows =
     rowsPerPage - Math.min(rowsPerPage, bucketRows.length - page * rowsPerPage);
-  const [open, setOpen] = React.useState(false);
   const theme = useTheme();
   const fullScreen = useMediaQuery(theme.breakpoints.down('sm'));
-
-  const handleClickOpen = () => {
-    setOpen(true);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
   return (
     <Paper>
       <TableContainer>
-          <div>
-            <ToastContainer onClick={() => handleClickOpen()}/>
-            <Button variant="outlined" color="primary" onClick={handleClickOpen}>
-              Open responsive dialog
-            </Button>
-            <Dialog
-                fullScreen={fullScreen}
-                open={open}
-                onClose={handleClose}
-                aria-labelledby="responsive-dialog-title"
-              >
-                <DialogTitle id="responsive-dialog-title">{"Downdload these files?"}</DialogTitle>
-                <DialogContent>
-                  <DialogContentText>
-                    <ul>
-                      <li>1</li>
-                      <li>2</li>
-                    </ul>
-                  </DialogContentText>
-                </DialogContent>
-                <DialogActions>
-                  <Button autoFocus onClick={handleClose} color="primary">
-                    Disagree
-                  </Button>
-                  <Button onClick={handleClose} color="primary" autoFocus>
-                    Agree
-                  </Button>
-                </DialogActions>
-              </Dialog>
-          </div>
           <Table>
           <EnhancedTableHead
             numSelected={selected.length}
@@ -695,157 +716,12 @@ const Browser = () => {
     setBucketItems(bucketItemRows);
   }, [bucketSelected]);
 
-  const handleOpen = (state) => {
-    console.log(permissions)
-    setOpen(state)
-  }
-
-  const handleGenerateKey= _ => {
-    setBucketKey("A883DA");
-  }
-  const [open, setOpen] = useState(false);
-  const [bucketKey, setBucketKey] = useState('000000')
-  const currentExpireDate = new Date();
-  const [permissionState, setPermissionState] = useState({
-    GetFileList: false,
-    GetFileListHidden: false,
-    Download: false,
-    DownloadHidden: false,
-    Upload: false,
-    UploadHidden: false,
-    DeleteFile: false, 
-    RecoverFile: false,
-  });
-  const { GetFileList, GetFileListHidden, Download, DownloadHidden, 
-    Upload, UploadHidden, DeleteFile, RecoverFile} = permissionState;
-
-  const handlePermissionChange  = (e) =>{
-    setPermissionState({...permissionState, [e.target.name]: e.target.checked})
-  }
-  
-  const permissions = [
-    {
-      value: GetFileList,
-      name: 'GetFileList',
-      label: 'Get file list',
-    },
-    {
-      value: GetFileListHidden,
-      name: 'GetFileListHidden',
-      label: 'Get hidden file list',
-    },
-    {
-      value: Download,
-      name: 'Download',
-      label: 'Download',
-    },
-    {
-      value: DownloadHidden,
-      name: 'DownloadHidden',
-      label: 'Download hidden',
-    },
-    {
-      value: Upload,
-      name: 'Upload',
-      label: 'Upload',
-    },
-    {
-      value: UploadHidden,
-      name: 'UploadHidden',
-      label: 'Upload hidden',
-    },
-    {
-      value: DeleteFile,
-      name: 'DeleteFile',
-      label: 'Delete file',
-    },
-    {
-      value: RecoverFile,
-      name: 'RecoverFile',
-      label: 'Recover file',
-    },
-  ];
   return (  
     <BucketContainer
       items={buckets}
       onItemClick={(name) => setBucketSelected(name)}
       visibility={bucketSelected !== null ? "hidden" : "visible"}
     >
-      <button type="button" onClick={() => handleOpen(true)}>
-        Open
-      </button>
-      {open ? (
-        <>
-          <div
-            className="justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none"
-          >
-            <div className="relative w-auto my-6 mx-auto max-w-3xl">
-              {/*content*/}
-              <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
-                {/*header*/}
-                <div className="flex items-start justify-between p-5 border-b border-solid border-gray-300 rounded-t">
-                  <h3 className="text-3xl font-semibold">
-                    Create new key
-                  </h3>
-                  <button
-                    className="p-1 ml-auto bg-transparent border-0 text-black opacity-5 float-right text-3xl leading-none font-semibold outline-none focus:outline-none"
-                    onClick={() => handleOpen(false)}
-                  >
-                    <span className="bg-transparent text-black opacity-5 h-6 w-6 text-2xl block outline-none focus:outline-none">
-                      ×
-                    </span>
-                  </button>
-                </div>
-                {/*body*/}
-                <div className="relative p-6 flex-auto">
-                  <FormControl >
-                    <FormLabel></FormLabel>
-                    <FormGroup className="grid grid-cols-2 grid-rows-4 gap-3">    
-                      {permissions.map((permission) => {
-                       return(<FormControlLabel control={<Checkbox checked={permission.value} onChange={handlePermissionChange} name={permission.name}/>} label={permission.label}>                         
-                       </FormControlLabel>)
-                      })}
-                    </FormGroup>
-                  </FormControl>
-                  <p className="my-4 text-gray-600 text-lg leading-relaxed">
-                    Choose key expiration date:
-                  </p>
-                  <form noValidate>
-                    <TextField
-                      id="expire-local"
-                      type="datetime-local"
-                      defaultValue={currentExpireDate}
-                      InputLabelProps={{
-                        shrink: true,
-                      }}
-                    />
-                  </form>
-                </div>
-                {/*footer*/}
-                <div className="flex items-center justify-end p-6 border-t border-solid border-gray-300 rounded-b">
-                  <button
-                    className="text-red-500 background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1"
-                    type="button"
-                    style={{ transition: "all .15s ease" }}
-                    onClick={() => handleOpen(false)}
-                  >
-                    Close
-                  </button>
-                  <button
-                    className="bg-light-blue text-white active:bg-light-blue font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1"
-                    type="button"
-                    style={{ transition: "all .15s ease" }}
-                    onClick={() => handleOpen(false)}
-                  >
-                    Create
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
-        </>
-      ) : null}
       <BucketItemsContainer
         title={bucketSelected}
         items={bucketItems}
